LIBITD-2455. Moved "TopTextbooksProcessor" into "textbooks/processor.py"
Refactored the "TopTextbooksProcessor" class out of the
"core/handling.py" class, into the "textbooks/processor.py" class.

This is part of a refactoring to have "generic" classes in "core", and
application-specific classes in "textbooks".

https://umd-dit.atlassian.net/browse/LIBITD-2455
