+++ tests/processor_test.py
@@ -1,4 +1,4 @@
-from core.handling import TopTextbooksProcessor
+from textbooks.processor import TopTextbooksProcessor
 
 
 def test_unique_mms_ids():
